+++ cmake/Config.cmake
@@ -1,0 +1,13 @@
+#PSVita App info
+set(VITA_APP_NAME "VitaMaterialKit")
+set(VITA_TITLEID "MKIT00001")
+set(VITA_VERSION "00.75")
+
+#Vita send
+set(PSVITA_IP "192.168.0.192")
+set(PSVITA_VOLUME "/Volumes/Untitled/")
+
+#Debug
+set(DEBUG_IP "192.168.0.108")
+set(DEBUG_PORT 18194)
+