+++ cmake/Files.cmake
@@ -1,0 +1,29 @@
+#Files
+set(FILES_SCE_SYS
+        sce_sys/icon0.png                       sce_sys/icon0.png
+        sce_sys/livearea/pic0.png               sce_sys/livearea/pic0.png
+        sce_sys/livearea/contents/bg.png        sce_sys/livearea/contents/bg.png
+        sce_sys/livearea/contents/startup.png   sce_sys/livearea/contents/startup.png
+        sce_sys/livearea/contents/template.xml  sce_sys/livearea/contents/template.xml
+        )
+
+set(FILES_FONTS
+        assets/fonts/mdi/materialdesignicons-webfont.ttf    assets/fonts/mdi/materialdesignicons-webfont.ttf
+        assets/fonts/Roboto/Roboto-Black.ttf                assets/fonts/Roboto/Roboto-Black.ttf
+        assets/fonts/Roboto/Roboto-BlackItalic.ttf          assets/fonts/Roboto/Roboto-BlackItalic.ttf
+        assets/fonts/Roboto/Roboto-Bold.ttf                 assets/fonts/Roboto/Roboto-Bold.ttf
+        assets/fonts/Roboto/Roboto-BoldItalic.ttf           assets/fonts/Roboto/Roboto-BoldItalic.ttf
+        assets/fonts/Roboto/Roboto-Italic.ttf               assets/fonts/Roboto/Roboto-Italic.ttf
+        assets/fonts/Roboto/Roboto-Light.ttf                assets/fonts/Roboto/Roboto-Light.ttf
+        assets/fonts/Roboto/Roboto-LightItalic.ttf          assets/fonts/Roboto/Roboto-LightItalic.ttf
+        assets/fonts/Roboto/Roboto-Medium.ttf               assets/fonts/Roboto/Roboto-Medium.ttf
+        assets/fonts/Roboto/Roboto-MediumItalic.ttf         assets/fonts/Roboto/Roboto-MediumItalic.ttf
+        assets/fonts/Roboto/Roboto-Regular.ttf              assets/fonts/Roboto/Roboto-Regular.ttf
+        assets/fonts/Roboto/Roboto-Thin.ttf                 assets/fonts/Roboto/Roboto-Thin.ttf
+        assets/fonts/Roboto/Roboto-ThinItalic.ttf           assets/fonts/Roboto/Roboto-ThinItalic.ttf
+        )
+
+set(FILES_LANGUAGES
+        i18n/en_us.json i18n/en_us.json
+        i18n/fr.json    i18n/fr.json
+        )