+++ cmake/Vita3K.cmake
@@ -1,0 +1,3 @@
+#Vita3K
+set(VITA3K_APP_PATH "/Applications/Vita3K.app/Contents/MacOS/Vita3K")
+set(VITA3K_FOLDER_PATH "/Users/Applelo/Library/Application\ Support/Vita3K/Vita3K/ux0/app/${VITA_TITLEID}")