+++ src/kit/core/App.cpp
@@ -22,7 +22,12 @@ void App::insertView(View *view) {
 
 void App::main() {
 
-    this->launchMountedViews();
+    this->beforeMountViews();
+    this->mountViews();
+
+    if (DEBUG_APP == 1) {
+        debugNetInit(DEBUG_IP, DEBUG_PORT, 3);
+    }
 
     while (run) {
         vita2d_start_drawing();
@@ -69,9 +74,13 @@ void App::checkExit() {
     if (viewsController->getActualView() == VIEWS_CONTROLLER_EXIT_LOOP) {
         run = 0;
     }
-    else if (viewsController->getActualView() == VIEWS_CONTROLLER_EXIT) {
+
+    else if (viewsController->getActualView() == VIEWS_CONTROLLER_EXIT_APP) {
         vita2d_fini();
         sceKernelExitProcess(0);
+        if (DEBUG_APP == 1) {
+            debugNetFinish();
+        }
     }
 }
 
@@ -104,7 +113,11 @@ void App::afterView() {
 
 }
 
-void App::launchMountedViews() {
+void App::beforeMountViews() {
+
+}
+
+void App::mountViews() {
     for(auto const &kv : views) {
         kv.second->mounted();
     }